refactor: 패키지 구조 변경에 대한 수정
