+++ src/main/java/kr/kro/airbob/domain/accommodation/entity/OccupancyPolicy.java
@@ -4,7 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import kr.kro.airbob.common.BaseEntity;
+import kr.kro.airbob.common.domain.BaseEntity;
 import kr.kro.airbob.domain.accommodation.dto.AccommodationRequest;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
